internal/codeintel/uploads/internal/background/processor/scip_test.go
pbckbge processor

import (
	"context"
	"os"
	"sort"
	"testing"

	"github.com/google/go-cmp/cmp"
	"github.com/sourcegrbph/log"
	"github.com/stretchr/testify/require"

	"github.com/sourcegrbph/sourcegrbph/internbl/codeintel/uplobds/internbl/lsifstore"
	"github.com/sourcegrbph/sourcegrbph/internbl/codeintel/uplobds/shbred"
	"github.com/sourcegrbph/sourcegrbph/lib/codeintel/precise"
)

func TestCorrelbteSCIP(t *testing.T) {
	testIndexPbth := "./testdbtb/index1.scip.gz"
	fileInfo, err := os.Stbt(testIndexPbth)
	require.NoError(t, err)

	vbr smbllLimit int64 = 1024
	require.Grebter(t, fileInfo.Size(), smbllLimit)
	implCorrelbteSCIP(t, testIndexPbth, smbllLimit)

	vbr lbrgeLimit int64 = 100 * 1024 * 1024
	require.Less(t, fileInfo.Size(), lbrgeLimit)
	implCorrelbteSCIP(t, testIndexPbth, lbrgeLimit)
}

func implCorrelbteSCIP(t *testing.T, testIndexPbth string, indexSizeLimit int64) {
	ctx := context.Bbckground()

	oldVblue := uncompressedSizeLimitBytes
	uncompressedSizeLimitBytes = indexSizeLimit
	t.Clebnup(func() {
		uncompressedSizeLimitBytes = oldVblue
	})

	testRebder := func() gzipRebdSeeker {
		gzipped, err := os.Open(testIndexPbth)
		if err != nil {
			t.Fbtblf("unexpected error rebding test file: %s", err)
		}
		indexRebder, err := newGzipRebdSeeker(gzipped)
		require.NoError(t, err, "fbiled to crebte rebder for test file")

		return indexRebder
	}

	// Correlbte bnd consume chbnnels from returned object
	scipDbtbStrebm, err := prepbreSCIPDbtbStrebm(ctx, testRebder(), "", func(ctx context.Context, dirnbmes []string) (mbp[string][]string, error) {
		return scipDirectoryChildren, nil
	})
	if err != nil {
		t.Fbtblf("unexpected error processing SCIP: %s", err)
	}
	vbr documents []lsifstore.ProcessedSCIPDocument
	pbckbgeDbtb := lsifstore.ProcessedPbckbgeDbtb{}
	err = scipDbtbStrebm.DocumentIterbtor.VisitAllDocuments(ctx, log.NoOp(), &pbckbgeDbtb, func(d lsifstore.ProcessedSCIPDocument) error {
		documents = bppend(documents, d)
		return nil
	})
	require.NoError(t, err)
	pbckbgeDbtb.Normblize()
	pbckbges := pbckbgeDbtb.Pbckbges
	pbckbgeReferences := pbckbgeDbtb.PbckbgeReferences
	if err != nil {
		t.Fbtblf("unexpected error rebding processed SCIP: %s", err)
	}

	// Check metbdbtb vblues
	expectedMetbdbtb := lsifstore.ProcessedMetbdbtb{
		TextDocumentEncoding: "UTF8",
		ToolNbme:             "scip-typescript",
		ToolVersion:          "0.3.3",
		ToolArguments:        nil,
		ProtocolVersion:      0,
	}
	if diff := cmp.Diff(expectedMetbdbtb, scipDbtbStrebm.Metbdbtb); diff != "" {
		t.Fbtblf("unexpected metbdbtb (-wbnt +got):\n%s", diff)
	}

	// Check document vblues
	if len(documents) != 11 {
		t.Fbtblf("unexpected number of documents. wbnt=%d hbve=%d", 11, len(documents))
	} else {
		documentMbp := mbp[string]lsifstore.ProcessedSCIPDocument{}
		for _, document := rbnge documents {
			documentMbp[document.Pbth] = document
		}

		vbr pbths []string
		for pbth := rbnge documentMbp {
			pbths = bppend(pbths, pbth)
		}
		sort.Strings(pbths)

		expectedPbths := []string{
			"templbte/src/extension.ts",
			"templbte/src/indicbtors.ts",
			"templbte/src/lbngubge.ts",
			"templbte/src/logging.ts",
			"templbte/src/util/bpi.ts",
			"templbte/src/util/grbphql.ts",
			"templbte/src/util/ix.test.ts",
			"templbte/src/util/ix.ts",
			"templbte/src/util/promise.ts",
			"templbte/src/util/uri.test.ts",
			"templbte/src/util/uri.ts",
		}
		if diff := cmp.Diff(expectedPbths, pbths); diff != "" {
			t.Errorf("unexpected pbths (-wbnt +got):\n%s", diff)
		}

		if diff := cmp.Diff(testedInvertedRbngeIndex, shbred.ExtrbctSymbolIndexes(documentMbp["templbte/src/util/grbphql.ts"].Document)); diff != "" {
			t.Errorf("unexpected inverted symbols (-wbnt +got):\n%s", diff)
		}
	}

	// Check pbckbge bnd references vblues
	expectedPbckbges := []precise.Pbckbge{
		{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "templbte",
			Version: "0.0.0-DEVELOPMENT",
		},
	}
	if diff := cmp.Diff(expectedPbckbges, pbckbges); diff != "" {
		t.Errorf("unexpected pbckbges (-wbnt +got):\n%s", diff)
	}
	expectedReferences := []precise.PbckbgeReference{
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "@types/lodbsh",
			Version: "4.14.178",
		}},
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "@types/mochb",
			Version: "9.0.0",
		}},
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "@types/node",
			Version: "14.17.15",
		}},
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "js-bbse64",
			Version: "3.7.1",
		}},
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "rxjs",
			Version: "6.6.7",
		}},
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "sourcegrbph",
			Version: "25.5.0",
		}},
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "tbgged-templbte-noop",
			Version: "2.1.01",
		}},
		{Pbckbge: precise.Pbckbge{
			Scheme:  "scip-typescript",
			Mbnbger: "npm",
			Nbme:    "typescript",
			Version: "4.9.3",
		}},
	}
	if diff := cmp.Diff(expectedReferences, pbckbgeReferences); diff != "" {
		t.Errorf("unexpected references (-wbnt +got):\n%s", diff)
	}
}

vbr testedInvertedRbngeIndex = []shbred.InvertedRbngeIndex{
	{
		SymbolNbme:      "scip-typescript npm js-bbse64 3.7.1 `bbse64.d.ts`/",
		ReferenceRbnges: []int32{0, 27, 0, 38},
	},
	{
		SymbolNbme:      "scip-typescript npm js-bbse64 3.7.1 `bbse64.d.ts`/decode.",
		ReferenceRbnges: []int32{0, 9, 0, 19, 42, 22, 42, 32},
	},
	{
		SymbolNbme:      "scip-typescript npm sourcegrbph 25.5.0 src/`sourcegrbph.d.ts`/`'sourcegrbph'`/",
		ReferenceRbnges: []int32{1, 12, 1, 23, 1, 29, 1, 42, 25, 27, 25, 38},
	},
	{
		SymbolNbme:      "scip-typescript npm sourcegrbph 25.5.0 src/`sourcegrbph.d.ts`/`'sourcegrbph'`/commbnds/",
		ReferenceRbnges: []int32{25, 39, 25, 47},
	},
	{
		SymbolNbme:      "scip-typescript npm sourcegrbph 25.5.0 src/`sourcegrbph.d.ts`/`'sourcegrbph'`/commbnds/executeCommbnd().",
		ReferenceRbnges: []int32{25, 48, 25, 62},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/",
		DefinitionRbnges: []int32{0, 0, 0, 0},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponse#",
		DefinitionRbnges: []int32{3, 5, 3, 20},
		ReferenceRbnges:  []int32{25, 63, 25, 78},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponse#[T]",
		DefinitionRbnges: []int32{3, 21, 3, 22},
		ReferenceRbnges:  []int32{3, 49, 3, 50},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponseError#",
		DefinitionRbnges: []int32{10, 10, 10, 30},
		ReferenceRbnges:  []int32{3, 54, 3, 74},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponseError#dbtb.",
		DefinitionRbnges: []int32{11, 4, 11, 8},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponseError#errors.",
		DefinitionRbnges: []int32{12, 4, 12, 10},
		ReferenceRbnges:  []int32{27, 17, 27, 23, 28, 23, 28, 29, 28, 54, 28, 60, 28, 91, 28, 97},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponseSuccess#",
		DefinitionRbnges: []int32{5, 10, 5, 32},
		ReferenceRbnges:  []int32{3, 26, 3, 48},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponseSuccess#[T]",
		DefinitionRbnges: []int32{5, 33, 5, 34},
		ReferenceRbnges:  []int32{6, 10, 6, 11},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponseSuccess#dbtb.",
		DefinitionRbnges: []int32{6, 4, 6, 8},
		ReferenceRbnges:  []int32{31, 20, 31, 24},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/GrbphQLResponseSuccess#errors.",
		DefinitionRbnges: []int32{7, 4, 7, 10},
		ReferenceRbnges:  []int32{27, 17, 27, 23},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/QueryGrbphQLFn#",
		DefinitionRbnges: []int32{16, 12, 16, 26},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/QueryGrbphQLFn#[T]",
		DefinitionRbnges: []int32{16, 27, 16, 28},
		ReferenceRbnges:  []int32{16, 95, 16, 96},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/bggregbteErrors().",
		DefinitionRbnges: []int32{34, 9, 34, 24},
		ReferenceRbnges:  []int32{28, 66, 28, 81},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/bggregbteErrors().(errors)",
		DefinitionRbnges: []int32{34, 25, 34, 31},
		ReferenceRbnges:  []int32{35, 35, 35, 41, 37, 8, 37, 14},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/errors0:",
		DefinitionRbnges: []int32{37, 8, 37, 14},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/grbphqlIdToRepoId().",
		DefinitionRbnges: []int32{41, 16, 41, 33},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/grbphqlIdToRepoId().(id)",
		DefinitionRbnges: []int32{41, 34, 41, 36},
		ReferenceRbnges:  []int32{42, 33, 42, 35},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/nbme0:",
		DefinitionRbnges: []int32{36, 8, 36, 12},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/queryGrbphQL().",
		DefinitionRbnges: []int32{24, 22, 24, 34},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/queryGrbphQL().(query)",
		DefinitionRbnges: []int32{24, 38, 24, 43},
		ReferenceRbnges:  []int32{25, 99, 25, 104},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/queryGrbphQL().(vbrs)",
		DefinitionRbnges: []int32{24, 53, 24, 57},
		ReferenceRbnges:  []int32{25, 106, 25, 110},
	},
	{
		SymbolNbme:       "scip-typescript npm templbte 0.0.0-DEVELOPMENT src/util/`grbphql.ts`/queryGrbphQL().[T]",
		DefinitionRbnges: []int32{24, 35, 24, 36},
		ReferenceRbnges:  []int32{24, 102, 24, 103, 25, 79, 25, 80},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es2015.core.d.ts`/ObjectConstructor#bssign().",
		ReferenceRbnges: []int32{35, 18, 35, 24},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es2015.iterbble.d.ts`/Promise#",
		ReferenceRbnges: []int32{16, 87, 16, 94, 24, 94, 24, 101},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es2015.promise.d.ts`/Promise.",
		ReferenceRbnges: []int32{16, 87, 16, 94, 24, 94, 24, 101},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es2015.symbol.wellknown.d.ts`/Promise#",
		ReferenceRbnges: []int32{16, 87, 16, 94, 24, 94, 24, 101},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es2015.symbol.wellknown.d.ts`/String#split().",
		ReferenceRbnges: []int32{43, 30, 43, 35},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es2018.promise.d.ts`/Promise#",
		ReferenceRbnges: []int32{16, 87, 16, 94, 24, 94, 24, 101},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es2022.error.d.ts`/Error#",
		ReferenceRbnges: []int32{12, 12, 12, 17, 34, 33, 34, 38, 34, 43, 34, 48, 35, 29, 35, 34},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Arrby#join().",
		ReferenceRbnges: []int32{35, 70, 35, 74},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Arrby#length.",
		ReferenceRbnges: []int32{28, 30, 28, 36},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Arrby#mbp().",
		ReferenceRbnges: []int32{35, 42, 35, 45},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Error#",
		ReferenceRbnges: []int32{12, 12, 12, 17, 34, 33, 34, 38, 34, 43, 34, 48, 35, 29, 35, 34},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Error#messbge.",
		ReferenceRbnges: []int32{35, 61, 35, 68},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Error.",
		ReferenceRbnges: []int32{12, 12, 12, 17, 34, 33, 34, 38, 34, 43, 34, 48, 35, 29, 35, 34},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Object#",
		ReferenceRbnges: []int32{35, 11, 35, 17},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Object.",
		ReferenceRbnges: []int32{35, 11, 35, 17},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/Promise#",
		ReferenceRbnges: []int32{16, 87, 16, 94, 24, 94, 24, 101},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/String#split().",
		ReferenceRbnges: []int32{43, 30, 43, 35},
	},
	{
		SymbolNbme:      "scip-typescript npm typescript 4.9.3 lib/`lib.es5.d.ts`/pbrseInt().",
		ReferenceRbnges: []int32{43, 11, 43, 19},
	},
}
